Allow for negative follow me height

+++ sw/airborne/modules/ctrl/follow_me.h
@@ -38,7 +38,7 @@
 extern uint8_t follow_me_distance;
 extern uint8_t follow_me_distance_2;
 extern uint8_t stdby_distance;
-extern uint8_t follow_me_height;
+extern int16_t follow_me_height;
 extern uint16_t follow_me_region;
 extern float follow_me_heading;
 extern float lateral_offset;
